go back to installing everything (on-the-fly tool install has issues currently), but break into two build layers which might speed up the pull?

+++ docker/install-viral-ngs.sh
@@ -1,29 +1,30 @@
 #!/bin/bash
 #
-# This script requires INSTALL_PATH (typically /opt/viral-ngs)
-# and VIRAL_NGS_PATH (typically /opt/viral-ngs/source) to be set.
+# This script requires INSTALL_PATH (typically /opt/viral-ngs),
+# VIRAL_NGS_PATH (typically /opt/viral-ngs/source), and
+# CONDA_DEFAULT_ENV (typically /opt/miniconda) to be set.
 #
-# A miniconda install must exist at /opt/miniconda
+# A miniconda install must exist at $CONDA_DEFAULT_ENV
+# and $CONDA_DEFAULT_ENV/bin must be in the PATH
 #
 # Otherwise, this only requires the existence of the following files:
-#    easy-deploy-script/easy-deploy-viral-ngs.sh
-#    requirements-conda.txt
-#    requirements-conda-tests.txt
-#    requirements-py3.txt
+#	requirements-minimal.txt
+#	requirements-conda.txt
+#	requirements-conda-tests.txt
+#	requirements-py3.txt
 
 set -e -o pipefail
 
-export VIRAL_CONDA_ENV_PATH=/opt/miniconda
-
 mkdir -p $INSTALL_PATH/viral-ngs-etc
-#mkdir -p $VIRAL_NGS_PATH/.git  # this is needed to make the setup script know we have/will have a git checkout
-ln -s $VIRAL_NGS_PATH $INSTALL_PATH/viral-ngs-etc/viral-ngs
-ln -s $VIRAL_CONDA_ENV_PATH $INSTALL_PATH/viral-ngs-etc/conda-env
-#ln $VIRAL_NGS_PATH/easy-deploy-script/easy-deploy-viral-ngs.sh $INSTALL_PATH
+if [ ! -f $INSTALL_PATH/viral-ngs-etc/viral-ngs ]; then
+	ln -s $VIRAL_NGS_PATH $INSTALL_PATH/viral-ngs-etc/viral-ngs
+fi
+if [ ! -f $INSTALL_PATH/viral-ngs-etc/conda-env ]; then
+	ln -s $CONDA_DEFAULT_ENV $INSTALL_PATH/viral-ngs-etc/conda-env
+fi
 
 # setup/install viral-ngs directory tree and conda dependencies
 sync
-#$INSTALL_PATH/easy-deploy-viral-ngs.sh setup-git-local
 
 # manually install it ourselves instead of using easy-deploy
 if [[ "$1" == "minimal" ]]; then
